doc_qa_agent/requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
python-docx
openai

